move config file parsing in separate file
imx_sdp.c is already large enouth, create a separate compile
unit for configuration file parsing.
