+++ imx_sdp.h
@@ -108,15 +108,7 @@ struct sdp_command {
 };
 #pragma pack ()
 
-int get_val(const char** pp, int base);
-const char *move_string(char *dest, const char *src, unsigned cnt);
 void dump_bytes(unsigned char *src, unsigned cnt, unsigned addr);
-
-char const *get_global_conf_path(void);
-char const *get_base_path(char const *argv0);
-char const *conf_file_name(char const *file, char const *base_path, char const *conf_path);
-struct sdp_dev *parse_conf(const char *filename);
-struct sdp_work *parse_cmd_args(int argc, char * const *argv);
 
 void perform_mem_work(struct sdp_dev *dev, struct mem_work *mem);
 int do_status(struct sdp_dev *dev);
