+++ imx_uart.c
@@ -40,6 +40,7 @@
 #include "portable.h"
 #include "imx_sdp.h"
 #include "imx_loader.h"
+#include "imx_loader_config.h"
 
 extern int debugmode;
